+++ Modules/Movment/Database/Migrations/2023_11_01_151016_create_antecedents_table.php
@@ -23,14 +23,19 @@ return new class extends Migration
             $table->unsignedBigInteger("movments_id")->nullable();
 
             $table->unsignedBigInteger("patients_id");
-            $table->foreign('patients_id')
-                ->references('id')
-                ->on('patients')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
 
             $table->timestamps();
         });
+
+        if (Schema::hasTable('antecedents') && Schema::hasTable('patients')) {
+            Schema::table('antecedents', function (Blueprint $table) {
+                $table->foreign('patients_id')
+                    ->references('id')
+                    ->on('patients')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
